Rename AddPost page to AddPostPage
