+++ src/router/index.tsx
@@ -6,7 +6,7 @@ import {
 } from "react-router-dom";
 import BlogPage from "../pages/BlogPage/BlogPage";
 import App from "../components/App/App";
-import AddPost from "../pages/AddPost/AddPost";
+import AddPost from "../pages/AddPostPage/AddPostPage";
 import PageNotFound from "../pages/PageNotFound/PageNotFound";
 
 const router = createBrowserRouter(
